Reorganization
This update reorganized the order of the twitter fields.  It grouped them up to make it easier for the user, and also added blank columns for sentiment analysis use using the Microsoft Azure Machine Learning add-in.  Also added more comments for user if they look at the code.  Numbered each of the titles to correspond with the appropriate code entry for twitter to make it easier to understand.